requirements.txt
langchain
pymilvus
pymupdf
sentence-transformers
psycopg2-binary
fastapi
uvicorn